+++ src/test/transaction_tests.cpp
@@ -482,7 +482,7 @@ BOOST_AUTO_TEST_CASE(test_big_transaction) {
     for (size_t i = 0; i < mtx.vin.size(); i++) {
         std::vector<CScriptCheck> vChecks;
         const CTxOut& out = coins[tx.vin[i].prevout.GetN()].GetTxOut();
-        vChecks.emplace_back(testConfig, true, out.scriptPubKey, out.nValue, tx, i,
+        vChecks.emplace_back(testConfig, true, out.scriptPubKey, out.nValue, tx, static_cast<unsigned int>(i),
                              MANDATORY_SCRIPT_VERIFY_FLAGS, false, txdata);
         
         control.Add(vChecks);
